agent/requirements.txt
mamba==0.8.3
msgpack-python==0.4.4
nanomsg==1.0
paramiko==1.15.2
